+++ shared/shared/src/main/scala/sc4pac/shared.scala
@@ -154,6 +154,7 @@ abstract class SharedData {
     variantDescriptions: Map[String, Map[String, String]] = Map.empty,  // variantKey -> variantValue -> description
     metadataSource: Option[SubPath] = None,  // path to yaml file
     metadataSourceUrl: Option[Uri] = None,  // full URL to yaml file
+    metadataIssueUrl: Option[Uri] = None,  // URL to create new issue for this package
     channelLabel: Option[String] = None,
   ) extends PackageAsset {
 
@@ -176,9 +177,13 @@ abstract class SharedData {
     author: String = "",
     images: Seq[String] = Seq.empty,
     website: String = "",
+    websites: Seq[String] = Seq.empty,
     requiredBy: Seq[BareModule] = Seq.empty,  // optional and only informative (mangles all variants and versions, is limited to one channel,
                                               // can easily become outdated since json files are cached indefinitely)
-  ) derives ReadWriter
+  ) derives ReadWriter {
+    def upgradeWebsites: Info =
+      if (websites.isEmpty && website.nonEmpty) copy(website = "", websites = Seq(website)) else this
+  }
   object Info {
     val empty = Info()
   }
@@ -266,11 +271,7 @@ abstract class SharedData {
     )
 
     private def findExternalIds(pkg: Package): Map[String, Seq[String]] = {
-      findExternalId(url = pkg.info.website) match {
-        // TODO currently there is just one website, but there could be multiple in the future
-        case Some(exchangeKey -> externalId) => Map(exchangeKey -> Seq(externalId))
-        case None => Map.empty
-      }
+      pkg.info.websites.flatMap(findExternalId).groupMap(_._1)(_._2)
     }
 
     def findExternalId(url: String): Option[(String, String)] = {  // stex/sc4e -> id
@@ -332,9 +333,10 @@ abstract class SharedData {
     case class Info(
       channelLabel: Option[String],
       metadataSourceUrl: Option[Uri],
+      metadataIssueUrl: Option[Uri] = None,
     ) derives ReadWriter
     object Info {
-      val empty = Info(None, None)
+      val empty = Info(None, None, None)
     }
 
     case class ExtPkg(group: String, name: String, checksum: Checksum) derives ReadWriter {